pyproject.toml
[project]
name = "mas-consensus-wkx"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "networkx[default]>=3.4.2",
    "numpy==1.26.4",
    "openai>=2.3.0",
    "pillow>=11.3.0",
    "scipy>=1.15.3",
    "seaborn>=0.13.2",
    "tqdm>=4.67.1",
]
